+++ app/src/pages/spaces/projects/ProjectTestcaseInfoPage/TestcaseNavigator/TestcaseNavigatorControl.scss
@@ -1,5 +1,6 @@
 .testcase-navigator-control-wrapper {
   $navigator-padding: 1rem;
+
   position: relative;
 
   .testcase-manage-button {
@@ -213,51 +214,6 @@
           display: none;
         }
       }
-    }
-  }
-
-  .testcase-filter-group {
-    &.floating {
-      position: absolute;
-      width: max(100%, 360px);
-      top: 60px;
-      left: 0px;
-      background-color: white;
-      border: 1px solid var(--border-light-color);
-      border-radius: 10px;
-      box-shadow: 10px 10px 30px rgba(0, 0, 0, 0.1);
-      z-index: 99;
-    }
-
-    padding: $navigator-padding;
-    border-bottom: 1px solid rgba(0, 0, 0, 0.12);
-    border-bottom: 1px solid var(--border-light-color);
-
-    display: flex;
-    flex-direction: column;
-    gap: 1rem 0;
-
-    .testcase-filter-ids-list,
-    .testcase-filter-release-ids-list {
-      overflow-x: auto;
-      display: flex;
-      flex-direction: row;
-
-      min-height: $sm-height;
-      margin-left: 90px;
-
-      & > div {
-        white-space: nowrap;
-      }
-    }
-  }
-  .filter-input-group {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    gap: 0 5px;
-    .selector-wrapper {
-      flex: 1;
     }
   }
 }
